+++ dcm_common/services/extensions/orchestration.py
@@ -13,7 +13,7 @@ from .common import (
     _ExtensionRequirement,
     ExtensionConditionRequirement,
 )
-from .notification import _connected
+from .notification import setup_connection_test_callback_with_state
 
 
 def _startup(
@@ -28,10 +28,9 @@ def _startup(
     requirements are met) or stops orchestrator if requirements are no
     longer met.
     """
-    first_try = orchestrator.running
     if orchestrator.running:
         if _ExtensionRequirement.check_requirements(
-            requirements, "Orchestrator startup delayed until '{}' is ready."
+            requirements, "Detected missing orchestrator-requirement '{}'."
         ):
             return
         orchestrator.stop(True)
@@ -43,7 +42,6 @@ def _startup(
             requirements, "Orchestrator startup delayed until '{}' is ready."
         ):
             break
-        first_try = False
         abort.wait(config.ORCHESTRATION_ABORT_NOTIFICATIONS_STARTUP_INTERVAL)
         if abort.is_set():
             return
@@ -52,8 +50,7 @@ def _startup(
         kwargs["cwd"] = config.FS_MOUNT_POINT
     orchestrator.run(**kwargs)
     result.ready.set()
-    if not first_try:
-        print_status("Orchestrator is ready.")
+    print_status("Orchestrator is ready.")
 
 
 # FIXME: drop legacy support
@@ -67,6 +64,8 @@ def orchestration(app, config, orchestrator, name, as_process) -> CDaemon:
     daemon is executed directly, i.e., in the same process from which
     this process has been called.
     """
+    _connected = setup_connection_test_callback_with_state(False)
+
     return orchestration_loader(
         app,
         config,